pyproject.toml
[tool.poetry]
name = "oascomply"
version = "0.1.0"
description = "OpenAPI Specification Compliance Tools"
authors = ["Henry H. Andrews <andrews_henry@yahoo.com>"]
readme = "README.md"

[tool.poetry.scripts]
oascomply = 'oascomply.apidescription:ApiDescription.load'
patch-oas-schemas = 'oascomply.patch:patch'

[tool.poetry.dependencies]
python = "^3.8"
jschon = "^0.10.2"
pyyaml = "^6.0"
rdflib = "^6.3.2"
pyshacl = "^0.22.1"
json-merge-patch = "^0.2"
abnf = "^2.2.0"
rfc3987 = "^1.3.8"
rfc3986 = {version = "^2.0.0", extras = ["idna"]}
urnparse = "^0.2.0"
json-source-map = "^1.0.5"
yaml-source-map = { git = "https://github.com/handrews/yaml-source-map.git", branch = "versions" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
